+++ game-likeadragonpirateyakuzainhawaii/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## FUTURE
+
+- Add: Xbox version support - NOTE: Xbox currently crashes on launch with SRMM injector installed.
+
+## [0.2.0] - 2026-04-17
+
+- Added: Root files installer
+- Fixed: Updated SRMM Mod installer to look for mod-meta.yaml file.
+
 ## [0.1.4] - 2025-03-12
 
 - Removed an obsolete parameter that caused downloading updates from the Moddding Tools domain, including SRMM, to fail with wrong URL